+++ js/app/src/Blocks.svelte
@@ -328,12 +328,15 @@
 
 		var a = target.getElementsByTagName("a");
 
+		const is_external_url = (link: string | null) =>
+			link && new URL(link).origin !== location.origin;
+
 		for (var i = 0; i < a.length; i++) {
 			const _target = a[i].getAttribute("target");
 			const _link = a[i].getAttribute("href");
 
-			// only target anchor tags that are not internal links
-			if (!_link?.startsWith("#") && _target !== "_blank")
+			// only target anchor tags with external links
+			if (is_external_url(_link) && _target !== "_blank")
 				a[i].setAttribute("target", "_blank");
 		}